engine structure: move fonts related sources into crengine/src/lvfont. crengine/src/lvfont: simpilfy `#include` statements.
