+++ android/app/CMakeLists.txt
@@ -84,11 +84,7 @@ set(CRENGINE_SRC_FILES
     ${CR3_ROOT}/crengine/src/txtselector.cpp
     ${CR3_ROOT}/crengine/src/crtest.cpp
     ${CR3_ROOT}/crengine/src/lvbmpbuf.cpp
-    ${CR3_ROOT}/crengine/src/lvfnt.cpp
     ${CR3_ROOT}/crengine/src/hyphman.cpp
-    ${CR3_ROOT}/crengine/src/lvfont.cpp
-    ${CR3_ROOT}/crengine/src/lvembeddedfont.cpp
-    ${CR3_ROOT}/crengine/src/lvfntman.cpp
     ${CR3_ROOT}/crengine/src/lvimg.cpp
     ${CR3_ROOT}/crengine/src/crskin.cpp
     ${CR3_ROOT}/crengine/src/lvdrawbuf.cpp
@@ -101,12 +97,16 @@ set(CRENGINE_SRC_FILES
     ${CR3_ROOT}/crengine/src/hist.cpp
     ${CR3_ROOT}/crengine/src/xxhash.c
     ${CR3_ROOT}/crengine/src/textlang.cpp
-    ${CR3_ROOT}/crengine/src/private/lvfontglyphcache.cpp
-    ${CR3_ROOT}/crengine/src/private/lvfontboldtransform.cpp
-    ${CR3_ROOT}/crengine/src/private/lvfontcache.cpp
-    ${CR3_ROOT}/crengine/src/private/lvfontdef.cpp
-    ${CR3_ROOT}/crengine/src/private/lvfreetypeface.cpp
-    ${CR3_ROOT}/crengine/src/private/lvfreetypefontman.cpp
+    ${CR3_ROOT}/crengine/src/lvfont/lvfnt.cpp
+    ${CR3_ROOT}/crengine/src/lvfont/lvembeddedfont.cpp
+    ${CR3_ROOT}/crengine/src/lvfont/lvfont.cpp
+    ${CR3_ROOT}/crengine/src/lvfont/lvfntman.cpp
+    ${CR3_ROOT}/crengine/src/lvfont/lvfontglyphcache.cpp
+    ${CR3_ROOT}/crengine/src/lvfont/lvfontboldtransform.cpp
+    ${CR3_ROOT}/crengine/src/lvfont/lvfontcache.cpp
+    ${CR3_ROOT}/crengine/src/lvfont/lvfontdef.cpp
+    ${CR3_ROOT}/crengine/src/lvfont/lvfreetypeface.cpp
+    ${CR3_ROOT}/crengine/src/lvfont/lvfreetypefontman.cpp
     ${CR3_ROOT}/crengine/fc-lang/fc-lang-cat.c
 )
 #    ${CR3_ROOT}/crengine/src/cri18n.cpp
