+++ crengine/CMakeLists.txt
@@ -33,11 +33,7 @@ SET (CRENGINE_SOURCES
 if ( NOT ${GUI} STREQUAL FB2PROPS )
     SET (CRENGINE_SOURCES ${CRENGINE_SOURCES}
         src/lvbmpbuf.cpp   
-        src/lvfnt.cpp      
         src/hyphman.cpp  
-        src/lvembeddedfont.cpp
-        src/lvfont.cpp
-        src/lvfntman.cpp        
         src/crgui.cpp     
         src/lvimg.cpp           
         src/crskin.cpp    
@@ -58,17 +54,21 @@ if ( NOT ${GUI} STREQUAL FB2PROPS )
         src/odtfmt.cpp
         src/odxutil.cpp
         src/crconcurrent.cpp
-        src/private/lvbasefont.cpp
-        src/private/lvbitmapfont.cpp
-        src/private/lvbitmapfontman.cpp
-        src/private/lvfontglyphcache.cpp
-        src/private/lvfontcache.cpp
-        src/private/lvfontboldtransform.cpp
-        src/private/lvfontdef.cpp
-        src/private/lvwin32font.cpp
-        src/private/lvwin32fontman.cpp
-        src/private/lvfreetypeface.cpp
-        src/private/lvfreetypefontman.cpp
+        src/lvfont/lvfont.cpp
+        src/lvfont/lvfntman.cpp
+        src/lvfont/lvfnt.cpp
+        src/lvfont/lvembeddedfont.cpp
+        src/lvfont/lvbasefont.cpp
+        src/lvfont/lvbitmapfont.cpp
+        src/lvfont/lvbitmapfontman.cpp
+        src/lvfont/lvfontglyphcache.cpp
+        src/lvfont/lvfontcache.cpp
+        src/lvfont/lvfontboldtransform.cpp
+        src/lvfont/lvfontdef.cpp
+        src/lvfont/lvwin32font.cpp
+        src/lvfont/lvwin32fontman.cpp
+        src/lvfont/lvfreetypeface.cpp
+        src/lvfont/lvfreetypefontman.cpp
         #src/xutils.cpp
     )
 endif (NOT ${GUI} STREQUAL FB2PROPS)
