+++ crengine/Tools/GammaGen/gen.sh.cmake
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-@CMAKE_CURRENT_BINARY_DIR@/gammagen@CMAKE_EXECUTABLE_SUFFIX@ @CMAKE_SOURCE_DIR@/crengine/include/gammatbl.h
+@CMAKE_CURRENT_BINARY_DIR@/gammagen@CMAKE_EXECUTABLE_SUFFIX@ @CMAKE_SOURCE_DIR@/crengine/src/lvfont/gammatbl.h
