+++ crengine/src/lvfont/gammatbl.h
