+++ crengine/src/lvfont/lvbasefont.cpp
