+++ crengine/src/lvfont/lvbitmapfont.cpp
