+++ crengine/src/lvfont/lvbitmapfont.h
