+++ crengine/src/lvfont/lvbitmapfontman.cpp
