+++ crengine/src/lvfont/lvbitmapfontman.h
@@ -15,8 +15,8 @@
 #ifndef __LV_BITMAPFONTMAN_H_INCLUDED__
 #define __LV_BITMAPFONTMAN_H_INCLUDED__
 
-#include "../../include/crsetup.h"
-#include "../../include/lvfntman.h"
+#include "crsetup.h"
+#include "lvfntman.h"
 #include "lvfontcache.h"
