+++ crengine/src/lvfont/lvfnt.cpp
@@ -14,7 +14,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
-#include "../include/lvfnt.h"
+#include "lvfnt.h"
 
 #define MIN_FONT_SIZE 2048
 #define MAX_FONT_SIZE 0x100000
