+++ crengine/src/lvfont/lvfontboldtransform.cpp
