+++ crengine/src/lvfont/lvfontcache.cpp
@@ -13,8 +13,8 @@
 */
 
 #include "lvfontcache.h"
-#include "../../include/lvstyles.h"
-#include "../../include/crlog.h"
+#include "lvstyles.h"
+#include "crlog.h"
 
 
 LVFontCacheItem *LVFontCache::findDuplicate(const LVFontDef *def) {
