+++ crengine/src/lvfont/lvfontcache.h
@@ -15,10 +15,10 @@
 #ifndef __LV_FONTCACHE_H_INCLUDED__
 #define __LV_FONTCACHE_H_INCLUDED__
 
-#include "../../include/crsetup.h"
-#include "../../include/lvfont.h"
-#include "../../include/lvptrvec.h"
-#include "../../include/lvstring32collection.h"
+#include "crsetup.h"
+#include "lvfont.h"
+#include "lvptrvec.h"
+#include "lvstring32collection.h"
 #include "lvfontdef.h"
 
 /// font cache item
