+++ crengine/src/lvfont/lvfontdef.cpp
