+++ crengine/src/lvfont/lvfontglyphcache.cpp
