+++ crengine/src/lvfont/lvfreetypeface.cpp
@@ -14,16 +14,16 @@
 
 #include "lvfreetypeface.h"
 
-#include "../../include/lvfntman.h"
-#include "../../include/lvfnt.h"
-#include "../../include/lvtextfm.h"
-#include "../../include/crlog.h"
+#include "lvfntman.h"
+#include "lvfnt.h"
+#include "lvtextfm.h"
+#include "crlog.h"
 #include "lvfontglyphcache.h"
 #include "lvfontdef.h"
 #include "lvfontcache.h"
 
 
-#include "../include/gammatbl.h"
+#include "gammatbl.h"
 
 
 // fc-lang database
