+++ crengine/src/lvfont/lvwin32fontman.cpp
